+++ TinyNvidiaUpdateChecker/DriverDialog.Designer.cs
@@ -1,0 +1,174 @@
+﻿using System.Windows.Forms;
+
+namespace TinyNvidiaUpdateChecker
+{
+    partial class DriverDialog
+    {
+        /// <summary>
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary>
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Windows Form Designer generated code
+
+        /// <summary>
+        /// Required method for Designer support - do not modify
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            System.ComponentModel.ComponentResourceManager resources = new System.ComponentModel.ComponentResourceManager(typeof(DriverDialog));
+            this.DownloadIntallButton = new System.Windows.Forms.Button();
+            this.DownloadBtn = new System.Windows.Forms.Button();
+            this.NotesBtn = new System.Windows.Forms.Button();
+            this.TitleLabel = new System.Windows.Forms.Label();
+            this.groupBox1 = new System.Windows.Forms.GroupBox();
+            this.ReleasedLabel = new System.Windows.Forms.Label();
+            this.VersionLabel = new System.Windows.Forms.Label();
+            this.IgnoreBtn = new System.Windows.Forms.Button();
+            this.webBrowser1 = new System.Windows.Forms.WebBrowser();
+            this.groupBox1.SuspendLayout();
+            this.SuspendLayout();
+            // 
+            // DownloadIntallButton
+            // 
+            this.DownloadIntallButton.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this.DownloadIntallButton.Location = new System.Drawing.Point(12, 214);
+            this.DownloadIntallButton.Name = "DownloadIntallButton";
+            this.DownloadIntallButton.Size = new System.Drawing.Size(90, 55);
+            this.DownloadIntallButton.TabIndex = 0;
+            this.DownloadIntallButton.Text = "Download + Install";
+            this.DownloadIntallButton.UseVisualStyleBackColor = true;
+            // 
+            // DownloadBtn
+            // 
+            this.DownloadBtn.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this.DownloadBtn.Location = new System.Drawing.Point(108, 214);
+            this.DownloadBtn.Name = "DownloadBtn";
+            this.DownloadBtn.Size = new System.Drawing.Size(89, 55);
+            this.DownloadBtn.TabIndex = 1;
+            this.DownloadBtn.Text = "Download";
+            this.DownloadBtn.UseVisualStyleBackColor = true;
+            // 
+            // NotesBtn
+            // 
+            this.NotesBtn.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this.NotesBtn.Location = new System.Drawing.Point(203, 214);
+            this.NotesBtn.Name = "NotesBtn";
+            this.NotesBtn.Size = new System.Drawing.Size(88, 55);
+            this.NotesBtn.TabIndex = 2;
+            this.NotesBtn.Text = "Release Notes";
+            this.NotesBtn.UseVisualStyleBackColor = true;
+            this.NotesBtn.Click += new System.EventHandler(this.NotesBtn_Click);
+            // 
+            // TitleLabel
+            // 
+            this.TitleLabel.AutoSize = true;
+            this.TitleLabel.Location = new System.Drawing.Point(9, 9);
+            this.TitleLabel.Name = "TitleLabel";
+            this.TitleLabel.Size = new System.Drawing.Size(260, 17);
+            this.TitleLabel.TabIndex = 5;
+            this.TitleLabel.Text = "New graphics card drivers are available!\r\n";
+            // 
+            // groupBox1
+            // 
+            this.groupBox1.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Top | System.Windows.Forms.AnchorStyles.Right)));
+            this.groupBox1.Controls.Add(this.ReleasedLabel);
+            this.groupBox1.Controls.Add(this.VersionLabel);
+            this.groupBox1.Location = new System.Drawing.Point(318, 72);
+            this.groupBox1.Name = "groupBox1";
+            this.groupBox1.Size = new System.Drawing.Size(200, 96);
+            this.groupBox1.TabIndex = 6;
+            this.groupBox1.TabStop = false;
+            // 
+            // ReleasedLabel
+            // 
+            this.ReleasedLabel.Location = new System.Drawing.Point(6, 49);
+            this.ReleasedLabel.Name = "ReleasedLabel";
+            this.ReleasedLabel.Size = new System.Drawing.Size(188, 44);
+            this.ReleasedLabel.TabIndex = 1;
+            this.ReleasedLabel.Text = "Released: ";
+            this.ReleasedLabel.TextAlign = System.Drawing.ContentAlignment.MiddleCenter;
+            // 
+            // VersionLabel
+            // 
+            this.VersionLabel.Location = new System.Drawing.Point(9, 15);
+            this.VersionLabel.Name = "VersionLabel";
+            this.VersionLabel.Size = new System.Drawing.Size(185, 44);
+            this.VersionLabel.TabIndex = 0;
+            this.VersionLabel.Text = "Version: ";
+            this.VersionLabel.TextAlign = System.Drawing.ContentAlignment.MiddleCenter;
+            // 
+            // IgnoreBtn
+            // 
+            this.IgnoreBtn.Anchor = ((System.Windows.Forms.AnchorStyles)((System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Left)));
+            this.IgnoreBtn.Location = new System.Drawing.Point(297, 214);
+            this.IgnoreBtn.Name = "IgnoreBtn";
+            this.IgnoreBtn.Size = new System.Drawing.Size(88, 55);
+            this.IgnoreBtn.TabIndex = 8;
+            this.IgnoreBtn.Text = "Ignore";
+            this.IgnoreBtn.UseVisualStyleBackColor = true;
+            this.IgnoreBtn.Click += new System.EventHandler(this.IgnoreBtn_Click);
+            // 
+            // webBrowser1
+            // 
+            this.webBrowser1.IsWebBrowserContextMenuEnabled = false;
+            this.webBrowser1.Location = new System.Drawing.Point(12, 29);
+            this.webBrowser1.MinimumSize = new System.Drawing.Size(20, 20);
+            this.webBrowser1.Name = "webBrowser1";
+            this.webBrowser1.ScriptErrorsSuppressed = true;
+            this.webBrowser1.Size = new System.Drawing.Size(300, 179);
+            this.webBrowser1.TabIndex = 9;
+            this.webBrowser1.DocumentCompleted += new System.Windows.Forms.WebBrowserDocumentCompletedEventHandler(this.webBrowser1_DocumentCompleted);
+            // 
+            // DriverDialog
+            // 
+            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.ClientSize = new System.Drawing.Size(530, 281);
+            this.Controls.Add(this.webBrowser1);
+            this.Controls.Add(this.IgnoreBtn);
+            this.Controls.Add(this.groupBox1);
+            this.Controls.Add(this.TitleLabel);
+            this.Controls.Add(this.NotesBtn);
+            this.Controls.Add(this.DownloadBtn);
+            this.Controls.Add(this.DownloadIntallButton);
+            this.FormBorderStyle = System.Windows.Forms.FormBorderStyle.None;
+            this.Icon = ((System.Drawing.Icon)(resources.GetObject("$this.Icon")));
+            this.Name = "DriverDialog";
+            this.StartPosition = System.Windows.Forms.FormStartPosition.CenterScreen;
+            this.Text = "Driver up!";
+            this.Load += new System.EventHandler(this.DriverDialog_Load);
+            this.groupBox1.ResumeLayout(false);
+            this.ResumeLayout(false);
+            this.PerformLayout();
+
+        }
+
+        #endregion
+
+        private System.Windows.Forms.Button DownloadIntallButton;
+        private System.Windows.Forms.Button DownloadBtn;
+        private System.Windows.Forms.Button NotesBtn;
+        private System.Windows.Forms.Label TitleLabel;
+        private System.Windows.Forms.GroupBox groupBox1;
+        private System.Windows.Forms.Label VersionLabel;
+        private System.Windows.Forms.Label ReleasedLabel;
+        private System.Windows.Forms.Button IgnoreBtn;
+        private System.Windows.Forms.WebBrowser webBrowser1;
+
+    }
+}